more contentScript messenger migrations
