+++ src/contentScript/messenger/api.ts
@@ -44,7 +44,6 @@ export const toggleQuickBar = getMethod("TOGGLE_QUICK_BAR");
 
 export const insertPanel = getMethod("INSERT_PANEL");
 export const insertButton = getMethod("INSERT_BUTTON");
-export const getAttributeExamples = getMethod("GET_ATTRIBUTE_EXAMPLES");
 
 export const runBlock = getMethod("RUN_SINGLE_BLOCK");
 export const runRendererBlock = getMethod("RUN_RENDERER_BLOCK");
@@ -66,9 +65,5 @@ export const runHeadlessPipeline = getMethod("RUN_HEADLESS_PIPELINE");
 export const runMapArgs = getMethod("RUN_MAP_ARGS");
 
 export const getCopilotHostData = getMethod("GET_COPILOT_HOST_DATA");
-
-export const reloadMarketplaceEnhancements = getMethod(
-  "RELOAD_MARKETPLACE_ENHANCEMENTS",
-);
 
 export const showLoginBanner = getMethod("SHOW_LOGIN_BANNER");
