+++ src/contentScript/messenger/strict/api.ts
@@ -55,3 +55,8 @@ export const notify = {
 };
 
 export const cancelSelect = getMethod("CANCEL_SELECT_ELEMENT");
+
+export const reloadMarketplaceEnhancements = getMethod(
+  "RELOAD_MARKETPLACE_ENHANCEMENTS",
+);
+export const getAttributeExamples = getMethod("GET_ATTRIBUTE_EXAMPLES");
