+++ src/extensionConsole/pages/UpdateBanner.tsx
@@ -17,7 +17,7 @@
 
 import React from "react";
 import { Button } from "react-bootstrap";
-import { getAvailableVersion } from "@/background/messenger/api";
+import { getAvailableVersion } from "@/background/messenger/strict/api";
 import reportError from "@/telemetry/reportError";
 import Banner from "@/components/banner/Banner";
 import { gt } from "semver";
@@ -36,7 +36,7 @@ export function useUpdateAvailable(): boolean {
     }
   }, []);
 
-  return updateAvailable;
+  return Boolean(updateAvailable);
 }
 
 const UpdateBanner: React.FunctionComponent = () => {
